Move YOLO folder ro ObjectDetection
