+++ src/ObjectDetector/YOLO/index.js
@@ -10,10 +10,10 @@ Heavily derived from https://github.com/ModelDepot/tfjs-yolo-tiny (ModelDepot: m
 */
 
 import * as tf from '@tensorflow/tfjs';
-import Video from '../utils/Video';
-import { imgToTensor } from '../utils/imageUtilities';
-import callCallback from '../utils/callcallback';
-import CLASS_NAMES from './../utils/COCO_CLASSES';
+import Video from '../../utils/Video';
+import { imgToTensor } from '../../utils/imageUtilities';
+import callCallback from '../../utils/callcallback';
+import CLASS_NAMES from '../../utils/COCO_CLASSES';
 
 import {
   nonMaxSuppression,
