+++ src/index.js
@@ -10,7 +10,7 @@ import soundClassifier from './SoundClassifier/';
 import KNNClassifier from './KNNClassifier/';
 import featureExtractor from './FeatureExtractor/';
 import word2vec from './Word2vec/';
-import YOLO from './YOLO';
+import YOLO from './ObjectDetector/YOLO';
 import ObjectDetector from './ObjectDetector';
 import poseNet from './PoseNet';
 import * as imageUtils from './utils/imageUtilities';
